refactor: changed dir name
