+++ src/core/bezier-spline.js
