+++ src/core/wave.js
